Add ability to parse paragraphs, changes to bold/italics
Headings following paragraphs without a blank line seperating them will
be considered part of the paragraph, for now.

Overlapping sections of bold and italic will likely result in problems
unless one is entirely contained in the other.

parse_md_text() no longer uses parse_md_text_text() directly and does
it's job itself, this is probably best within the other functions that
use it as well, so it can then be removed.